refactor: 상세화면 팝업 데이터를 나타내는 PopupDetail을 PopupInformation으로 이름 변경 및 데이터 구조 변경
